[cdc] Refactor kafka cdc test (#2672)
