+++ paimon-flink/paimon-flink-cdc/src/test/java/org/apache/paimon/flink/action/cdc/kafka/KafkaMaxwellSyncTableActionITCase.java
@@ -18,524 +18,71 @@
 
 package org.apache.paimon.flink.action.cdc.kafka;
 
-import org.apache.paimon.catalog.Identifier;
-import org.apache.paimon.table.AbstractFileStoreTable;
-import org.apache.paimon.table.FileStoreTable;
-import org.apache.paimon.types.DataType;
-import org.apache.paimon.types.DataTypes;
-import org.apache.paimon.types.RowType;
-
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
 
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
+/** IT cases for {@link KafkaSyncTableAction}. */
+public class KafkaMaxwellSyncTableActionITCase extends KafkaSyncTableActionITCase {
 
-import static org.apache.flink.streaming.connectors.kafka.table.KafkaConnectorOptions.SCAN_STARTUP_MODE;
-import static org.apache.flink.streaming.connectors.kafka.table.KafkaConnectorOptions.SCAN_STARTUP_SPECIFIC_OFFSETS;
-import static org.apache.flink.streaming.connectors.kafka.table.KafkaConnectorOptions.SCAN_STARTUP_TIMESTAMP_MILLIS;
-import static org.apache.flink.streaming.connectors.kafka.table.KafkaConnectorOptions.ScanStartupMode.EARLIEST_OFFSET;
-import static org.apache.flink.streaming.connectors.kafka.table.KafkaConnectorOptions.ScanStartupMode.GROUP_OFFSETS;
-import static org.apache.flink.streaming.connectors.kafka.table.KafkaConnectorOptions.ScanStartupMode.LATEST_OFFSET;
-import static org.apache.flink.streaming.connectors.kafka.table.KafkaConnectorOptions.ScanStartupMode.SPECIFIC_OFFSETS;
-import static org.apache.flink.streaming.connectors.kafka.table.KafkaConnectorOptions.ScanStartupMode.TIMESTAMP;
-import static org.apache.flink.streaming.connectors.kafka.table.KafkaConnectorOptions.TOPIC;
-import static org.apache.flink.streaming.connectors.kafka.table.KafkaConnectorOptions.VALUE_FORMAT;
-import static org.apache.paimon.testutils.assertj.AssertionUtils.anyCauseMatches;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-/** IT cases for {@link KafkaMaxwellSyncTableActionITCase}. */
-public class KafkaMaxwellSyncTableActionITCase extends KafkaActionITCaseBase {
+    private static final String MAXWELL = "maxwell";
 
     @Test
     @Timeout(60)
     public void testSchemaEvolution() throws Exception {
-        runSingleTableSchemaEvolution("schemaevolution");
-    }
-
-    private void runSingleTableSchemaEvolution(String sourceDir) throws Exception {
-        final String topic = "schema_evolution";
-        createTestTopic(topic, 1, 1);
-        // ---------- Write the maxwell json into Kafka -------------------
-        List<String> lines =
-                readLines(String.format("kafka/maxwell/table/%s/maxwell-data-1.txt", sourceDir));
-        try {
-            writeRecordsToKafka(topic, lines);
-        } catch (Exception e) {
-            throw new Exception("Failed to write maxwell data to Kafka.", e);
-        }
-        Map<String, String> kafkaConfig = getBasicKafkaConfig();
-        kafkaConfig.put(VALUE_FORMAT.key(), "maxwell-json");
-        kafkaConfig.put(TOPIC.key(), topic);
-        KafkaSyncTableAction action =
-                syncTableActionBuilder(kafkaConfig)
-                        .withPrimaryKeys("id")
-                        .withTableConfig(getBasicTableConfig())
-                        .build();
-        runActionWithDefaultEnv(action);
-
-        testSchemaEvolutionImpl(topic, sourceDir);
-    }
-
-    private void testSchemaEvolutionImpl(String topic, String sourceDir) throws Exception {
-        FileStoreTable table = getFileStoreTable(tableName);
-
-        RowType rowType =
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING().notNull(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING()
-                        },
-                        new String[] {"id", "name", "description", "weight"});
-        List<String> primaryKeys = Collections.singletonList("id");
-        List<String> expected =
-                Arrays.asList(
-                        "+I[101, scooter, Small 2-wheel scooter, 3.14]",
-                        "+I[102, car battery, 12V car battery, 8.1]");
-        waitForResult(expected, table, rowType, primaryKeys);
-
-        try {
-            writeRecordsToKafka(
-                    topic,
-                    readLines(
-                            String.format("kafka/maxwell/table/%s/maxwell-data-2.txt", sourceDir)));
-        } catch (Exception e) {
-            throw new Exception("Failed to write maxwell data to Kafka.", e);
-        }
-        rowType =
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING().notNull(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING()
-                        },
-                        new String[] {"id", "name", "description", "weight", "age"});
-        expected =
-                Arrays.asList(
-                        "+I[101, scooter, Small 2-wheel scooter, 3.14, NULL]",
-                        "+I[102, car battery, 12V car battery, 8.1, NULL]",
-                        "+I[103, 12-pack drill bits, 12-pack of drill bits with sizes ranging from #40 to #3, 0.8, 18]",
-                        "+I[104, hammer, 12oz carpenter's hammer, 0.75, 24]");
-        waitForResult(expected, table, rowType, primaryKeys);
-
-        try {
-            writeRecordsToKafka(
-                    topic,
-                    readLines(
-                            String.format("kafka/maxwell/table/%s/maxwell-data-3.txt", sourceDir)));
-        } catch (Exception e) {
-            throw new Exception("Failed to write maxwell data to Kafka.", e);
-        }
-        rowType =
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING().notNull(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING()
-                        },
-                        new String[] {"id", "name", "description", "weight", "age", "address"});
-        expected =
-                Arrays.asList(
-                        "+I[102, car battery, 12V car battery, 8.1, NULL, NULL]",
-                        "+I[103, 12-pack drill bits, 12-pack of drill bits with sizes ranging from #40 to #3, 0.8, 18, NULL]",
-                        "+I[104, hammer, 12oz carpenter's hammer, 0.75, 24, NULL]",
-                        "+I[105, hammer, 14oz carpenter's hammer, 0.875, NULL, Beijing]",
-                        "+I[107, rocks, box of assorted rocks, 5.3, NULL, NULL]");
-        waitForResult(expected, table, rowType, primaryKeys);
+        runSingleTableSchemaEvolution("schemaevolution", MAXWELL);
     }
 
     @Test
     @Timeout(60)
     public void testNotSupportFormat() throws Exception {
-        final String topic = "not_support";
-        createTestTopic(topic, 1, 1);
-        // ---------- Write the maxwell json into Kafka -------------------
-        List<String> lines = readLines("kafka/maxwell/table/schemaevolution/maxwell-data-1.txt");
-        try {
-            writeRecordsToKafka(topic, lines);
-        } catch (Exception e) {
-            throw new Exception("Failed to write maxwell data to Kafka.", e);
-        }
-        Map<String, String> kafkaConfig = getBasicKafkaConfig();
-        kafkaConfig.put(VALUE_FORMAT.key(), "togg-json");
-        kafkaConfig.put(TOPIC.key(), topic);
-        KafkaSyncTableAction action =
-                syncTableActionBuilder(kafkaConfig)
-                        .withPrimaryKeys("id")
-                        .withTableConfig(getBasicTableConfig())
-                        .build();
-
-        assertThatThrownBy(action::run)
-                .satisfies(
-                        anyCauseMatches(
-                                UnsupportedOperationException.class,
-                                "This format: togg-json is not supported."));
+        testNotSupportFormat(MAXWELL);
     }
 
     @Test
     @Timeout(60)
     public void testAssertSchemaCompatible() throws Exception {
-        final String topic = "assert_schema_compatible";
-        createTestTopic(topic, 1, 1);
-        // ---------- Write the maxwell json into Kafka -------------------
-        List<String> lines = readLines("kafka/maxwell/table/schemaevolution/maxwell-data-1.txt");
-        try {
-            writeRecordsToKafka(topic, lines);
-        } catch (Exception e) {
-            throw new Exception("Failed to write maxwell data to Kafka.", e);
-        }
-        Map<String, String> kafkaConfig = getBasicKafkaConfig();
-        kafkaConfig.put(VALUE_FORMAT.key(), "maxwell-json");
-        kafkaConfig.put(TOPIC.key(), topic);
-
-        // create an incompatible table
-        createFileStoreTable(
-                RowType.of(
-                        new DataType[] {DataTypes.STRING(), DataTypes.STRING()},
-                        new String[] {"k", "v1"}),
-                Collections.emptyList(),
-                Collections.singletonList("k"),
-                Collections.emptyMap());
-
-        KafkaSyncTableAction action =
-                syncTableActionBuilder(kafkaConfig)
-                        .withPrimaryKeys("id")
-                        .withTableConfig(getBasicTableConfig())
-                        .build();
-
-        assertThatThrownBy(action::run)
-                .satisfies(
-                        anyCauseMatches(
-                                IllegalArgumentException.class,
-                                "Paimon schema and source table schema are not compatible.\n"
-                                        + "Paimon fields are: [`k` STRING NOT NULL, `v1` STRING].\n"
-                                        + "Source table fields are: [`id` STRING NOT NULL, `name` STRING, `description` STRING, `weight` STRING]"));
+        testAssertSchemaCompatible(MAXWELL);
     }
 
     @Test
     @Timeout(60)
     public void testStarUpOptionSpecific() throws Exception {
-        final String topic = "start_up_specific";
-        createTestTopic(topic, 1, 1);
-        // ---------- Write the maxwell json into Kafka -------------------
-        List<String> lines = readLines("kafka/maxwell/table/startupmode/maxwell-data-1.txt");
-        try {
-            writeRecordsToKafka(topic, lines);
-        } catch (Exception e) {
-            throw new Exception("Failed to write maxwell data to Kafka.", e);
-        }
-        Map<String, String> kafkaConfig = getBasicKafkaConfig();
-        kafkaConfig.put(VALUE_FORMAT.key(), "maxwell-json");
-        kafkaConfig.put(TOPIC.key(), topic);
-        kafkaConfig.put(SCAN_STARTUP_MODE.key(), SPECIFIC_OFFSETS.toString());
-        kafkaConfig.put(SCAN_STARTUP_SPECIFIC_OFFSETS.key(), "partition:0,offset:1");
-        KafkaSyncTableAction action =
-                syncTableActionBuilder(kafkaConfig)
-                        .withPrimaryKeys("id")
-                        .withTableConfig(getBasicTableConfig())
-                        .build();
-        runActionWithDefaultEnv(action);
-
-        FileStoreTable table = getFileStoreTable(tableName);
-
-        RowType rowType =
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING().notNull(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING()
-                        },
-                        new String[] {"id", "name", "description", "weight"});
-        List<String> primaryKeys = Collections.singletonList("id");
-        // topic has two records we read two
-        List<String> expected =
-                Collections.singletonList("+I[102, car battery, 12V car battery, 8.1]");
-        waitForResult(expected, table, rowType, primaryKeys);
+        testStarUpOptionSpecific(MAXWELL);
     }
 
     @Test
     @Timeout(60)
     public void testStarUpOptionLatest() throws Exception {
-        final String topic = "start_up_latest";
-        createTestTopic(topic, 1, 1);
-        // ---------- Write the maxwell json into Kafka -------------------
-        List<String> lines = readLines("kafka/maxwell/table/startupmode/maxwell-data-1.txt");
-        try {
-            writeRecordsToKafka(topic, lines);
-        } catch (Exception e) {
-            throw new Exception("Failed to write maxwell data to Kafka.", e);
-        }
-        Map<String, String> kafkaConfig = getBasicKafkaConfig();
-        kafkaConfig.put(VALUE_FORMAT.key(), "maxwell-json");
-        kafkaConfig.put(TOPIC.key(), topic);
-        kafkaConfig.put(SCAN_STARTUP_MODE.key(), LATEST_OFFSET.toString());
-        KafkaSyncTableAction action =
-                syncTableActionBuilder(kafkaConfig)
-                        .withPrimaryKeys("id")
-                        .withTableConfig(getBasicTableConfig())
-                        .build();
-        runActionWithDefaultEnv(action);
-
-        Thread.sleep(5000);
-        FileStoreTable table = getFileStoreTable(tableName);
-        try {
-            writeRecordsToKafka(
-                    topic, readLines("kafka/maxwell/table/startupmode/maxwell-data-2.txt"));
-        } catch (Exception e) {
-            throw new Exception("Failed to write maxwell data to Kafka.", e);
-        }
-
-        RowType rowType =
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING().notNull(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING()
-                        },
-                        new String[] {"id", "name", "description", "weight"});
-        List<String> primaryKeys = Collections.singletonList("id");
-        // topic has four records we read two
-        List<String> expected =
-                Arrays.asList(
-                        "+I[103, 12-pack drill bits, 12-pack of drill bits with sizes ranging from #40 to #3, 0.8]",
-                        "+I[104, hammer, 12oz carpenter's hammer, 0.75]");
-        waitForResult(expected, table, rowType, primaryKeys);
+        testStarUpOptionLatest(MAXWELL);
     }
 
     @Test
     @Timeout(60)
     public void testStarUpOptionTimestamp() throws Exception {
-        final String topic = "start_up_timestamp";
-        createTestTopic(topic, 1, 1);
-        // ---------- Write the maxwell json into Kafka -------------------
-        List<String> lines = readLines("kafka/maxwell/table/startupmode/maxwell-data-1.txt");
-        try {
-            writeRecordsToKafka(topic, lines);
-        } catch (Exception e) {
-            throw new Exception("Failed to write maxwell data to Kafka.", e);
-        }
-        Map<String, String> kafkaConfig = getBasicKafkaConfig();
-        kafkaConfig.put(VALUE_FORMAT.key(), "maxwell-json");
-        kafkaConfig.put(TOPIC.key(), topic);
-        kafkaConfig.put(SCAN_STARTUP_MODE.key(), TIMESTAMP.toString());
-        kafkaConfig.put(
-                SCAN_STARTUP_TIMESTAMP_MILLIS.key(), String.valueOf(System.currentTimeMillis()));
-        KafkaSyncTableAction action =
-                syncTableActionBuilder(kafkaConfig)
-                        .withPrimaryKeys("id")
-                        .withTableConfig(getBasicTableConfig())
-                        .build();
-        runActionWithDefaultEnv(action);
-
-        try {
-            writeRecordsToKafka(
-                    topic, readLines("kafka/maxwell/table/startupmode/maxwell-data-2.txt"));
-        } catch (Exception e) {
-            throw new Exception("Failed to write maxwell data to Kafka.", e);
-        }
-        FileStoreTable table = getFileStoreTable(tableName);
-
-        RowType rowType =
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING().notNull(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING()
-                        },
-                        new String[] {"id", "name", "description", "weight"});
-        List<String> primaryKeys = Collections.singletonList("id");
-        // topic has four records we read two
-        List<String> expected =
-                Arrays.asList(
-                        "+I[103, 12-pack drill bits, 12-pack of drill bits with sizes ranging from #40 to #3, 0.8]",
-                        "+I[104, hammer, 12oz carpenter's hammer, 0.75]");
-        waitForResult(expected, table, rowType, primaryKeys);
+        testStarUpOptionTimestamp(MAXWELL);
     }
 
     @Test
     @Timeout(60)
     public void testStarUpOptionEarliest() throws Exception {
-        final String topic = "start_up_earliest";
-        createTestTopic(topic, 1, 1);
-        // ---------- Write the maxwell json into Kafka -------------------
-        List<String> lines = readLines("kafka/maxwell/table/startupmode/maxwell-data-1.txt");
-        try {
-            writeRecordsToKafka(topic, lines);
-        } catch (Exception e) {
-            throw new Exception("Failed to write maxwell data to Kafka.", e);
-        }
-        Map<String, String> kafkaConfig = getBasicKafkaConfig();
-        kafkaConfig.put(VALUE_FORMAT.key(), "maxwell-json");
-        kafkaConfig.put(TOPIC.key(), topic);
-        kafkaConfig.put(SCAN_STARTUP_MODE.key(), EARLIEST_OFFSET.toString());
-        KafkaSyncTableAction action =
-                syncTableActionBuilder(kafkaConfig)
-                        .withPrimaryKeys("id")
-                        .withTableConfig(getBasicTableConfig())
-                        .build();
-        runActionWithDefaultEnv(action);
-
-        try {
-            writeRecordsToKafka(
-                    topic, readLines("kafka/maxwell/table/startupmode/maxwell-data-2.txt"));
-        } catch (Exception e) {
-            throw new Exception("Failed to write maxwell data to Kafka.", e);
-        }
-        FileStoreTable table = getFileStoreTable(tableName);
-
-        RowType rowType =
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING().notNull(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING()
-                        },
-                        new String[] {"id", "name", "description", "weight"});
-        List<String> primaryKeys = Collections.singletonList("id");
-        // topic has four records we read all
-        List<String> expected =
-                Arrays.asList(
-                        "+I[101, scooter, Small 2-wheel scooter, 3.14]",
-                        "+I[102, car battery, 12V car battery, 8.1]",
-                        "+I[103, 12-pack drill bits, 12-pack of drill bits with sizes ranging from #40 to #3, 0.8]",
-                        "+I[104, hammer, 12oz carpenter's hammer, 0.75]");
-        waitForResult(expected, table, rowType, primaryKeys);
+        testStarUpOptionEarliest(MAXWELL);
     }
 
     @Test
     @Timeout(60)
     public void testStarUpOptionGroup() throws Exception {
-        final String topic = "start_up_group";
-        createTestTopic(topic, 1, 1);
-        // ---------- Write the maxwell json into Kafka -------------------
-        List<String> lines = readLines("kafka/maxwell/table/startupmode/maxwell-data-1.txt");
-        try {
-            writeRecordsToKafka(topic, lines);
-        } catch (Exception e) {
-            throw new Exception("Failed to write maxwell data to Kafka.", e);
-        }
-        Map<String, String> kafkaConfig = getBasicKafkaConfig();
-        kafkaConfig.put(VALUE_FORMAT.key(), "maxwell-json");
-        kafkaConfig.put(TOPIC.key(), topic);
-        kafkaConfig.put(SCAN_STARTUP_MODE.key(), GROUP_OFFSETS.toString());
-        KafkaSyncTableAction action =
-                syncTableActionBuilder(kafkaConfig)
-                        .withPrimaryKeys("id")
-                        .withTableConfig(getBasicTableConfig())
-                        .build();
-        runActionWithDefaultEnv(action);
-
-        try {
-            writeRecordsToKafka(
-                    topic, readLines("kafka/maxwell/table/startupmode/maxwell-data-2.txt"));
-        } catch (Exception e) {
-            throw new Exception("Failed to write maxwell data to Kafka.", e);
-        }
-        FileStoreTable table = getFileStoreTable(tableName);
-
-        RowType rowType =
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING().notNull(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING()
-                        },
-                        new String[] {"id", "name", "description", "weight"});
-        List<String> primaryKeys = Collections.singletonList("id");
-        // topic has four records we read all
-        List<String> expected =
-                Arrays.asList(
-                        "+I[101, scooter, Small 2-wheel scooter, 3.14]",
-                                "+I[102, car battery, 12V car battery, 8.1]",
-                        "+I[103, 12-pack drill bits, 12-pack of drill bits with sizes ranging from #40 to #3, 0.8]",
-                                "+I[104, hammer, 12oz carpenter's hammer, 0.75]");
-        waitForResult(expected, table, rowType, primaryKeys);
+        testStarUpOptionGroup(MAXWELL);
     }
 
     @Test
     @Timeout(60)
     public void testComputedColumn() throws Exception {
-        String topic = "computed_column";
-        createTestTopic(topic, 1, 1);
-
-        List<String> lines = readLines("kafka/maxwell/table/computedcolumn/maxwell-data-1.txt");
-        try {
-            writeRecordsToKafka(topic, lines);
-        } catch (Exception e) {
-            throw new Exception("Failed to write canal data to Kafka.", e);
-        }
-        Map<String, String> kafkaConfig = getBasicKafkaConfig();
-        kafkaConfig.put(VALUE_FORMAT.key(), "maxwell-json");
-        kafkaConfig.put(TOPIC.key(), topic);
-        KafkaSyncTableAction action =
-                syncTableActionBuilder(kafkaConfig)
-                        .withPartitionKeys("_year")
-                        .withPrimaryKeys("_id", "_year")
-                        .withComputedColumnArgs("_year=year(_date)")
-                        .withTableConfig(getBasicTableConfig())
-                        .build();
-        runActionWithDefaultEnv(action);
-
-        RowType rowType =
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING().notNull(),
-                            DataTypes.STRING(),
-                            DataTypes.INT().notNull()
-                        },
-                        new String[] {"_id", "_date", "_year"});
-        waitForResult(
-                Collections.singletonList("+I[101, 2023-03-23, 2023]"),
-                getFileStoreTable(tableName),
-                rowType,
-                Arrays.asList("_id", "_year"));
+        testComputedColumn(MAXWELL);
     }
 
     @Test
     @Timeout(60)
     public void testWaterMarkSyncTable() throws Exception {
-        String topic = "watermark";
-        createTestTopic(topic, 1, 1);
-        writeRecordsToKafka(topic, readLines("kafka/maxwell/table/watermark/maxwell-data-1.txt"));
-
-        Map<String, String> kafkaConfig = getBasicKafkaConfig();
-        kafkaConfig.put(VALUE_FORMAT.key(), "maxwell-json");
-        kafkaConfig.put(TOPIC.key(), topic);
-
-        Map<String, String> config = getBasicTableConfig();
-        config.put("tag.automatic-creation", "watermark");
-        config.put("tag.creation-period", "hourly");
-        config.put("scan.watermark.alignment.group", "alignment-group-1");
-        config.put("scan.watermark.alignment.max-drift", "20 s");
-        config.put("scan.watermark.alignment.update-interval", "1 s");
-
-        KafkaSyncTableAction action =
-                syncTableActionBuilder(kafkaConfig).withTableConfig(config).build();
-        runActionWithDefaultEnv(action);
-
-        AbstractFileStoreTable table =
-                (AbstractFileStoreTable) catalog.getTable(new Identifier(database, tableName));
-        while (true) {
-            if (table.snapshotManager().snapshotCount() > 0
-                    && table.snapshotManager().latestSnapshot().watermark()
-                            != -9223372036854775808L) {
-                return;
-            }
-            Thread.sleep(1000);
-        }
+        testWaterMarkSyncTable(MAXWELL);
     }
 }
