+++ paimon-flink/paimon-flink-cdc/src/test/java/org/apache/paimon/flink/action/cdc/kafka/KafkaOggSyncDatabaseActionITCase.java
@@ -18,498 +18,63 @@
 
 package org.apache.paimon.flink.action.cdc.kafka;
 
-import org.apache.paimon.catalog.FileSystemCatalogOptions;
-import org.apache.paimon.table.FileStoreTable;
-import org.apache.paimon.testutils.assertj.AssertionUtils;
-import org.apache.paimon.types.DataType;
-import org.apache.paimon.types.DataTypes;
-import org.apache.paimon.types.RowType;
-
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
 
-import javax.annotation.Nullable;
-
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.UUID;
-import java.util.concurrent.ThreadLocalRandom;
-
-import static org.apache.flink.streaming.connectors.kafka.table.KafkaConnectorOptions.TOPIC;
-import static org.apache.flink.streaming.connectors.kafka.table.KafkaConnectorOptions.VALUE_FORMAT;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
 /** IT cases for {@link KafkaSyncDatabaseAction}. */
-public class KafkaOggSyncDatabaseActionITCase extends KafkaActionITCaseBase {
+public class KafkaOggSyncDatabaseActionITCase extends KafkaSyncDatabaseActionITCase {
+    private static final String OGG = "ogg";
 
     @Test
     @Timeout(60)
     public void testSchemaEvolutionMultiTopic() throws Exception {
-        final String topic1 = "schema_evolution_0";
-        final String topic2 = "schema_evolution_1";
-        boolean writeOne = false;
-        int fileCount = 2;
-        List<String> topics = Arrays.asList(topic1, topic2);
-        topics.forEach(topic -> createTestTopic(topic, 1, 1));
-
-        // ---------- Write the ogg json into Kafka -------------------
-
-        for (int i = 0; i < fileCount; i++) {
-            try {
-                writeRecordsToKafka(
-                        topics.get(i),
-                        readLines(
-                                "kafka/ogg/database/schemaevolution/topic"
-                                        + i
-                                        + "/ogg-data-1.txt"));
-            } catch (Exception e) {
-                throw new Exception("Failed to write ogg data to Kafka.", e);
-            }
-        }
-
-        Map<String, String> kafkaConfig = getBasicKafkaConfig();
-        kafkaConfig.put(VALUE_FORMAT.key(), "ogg-json");
-        kafkaConfig.put(TOPIC.key(), String.join(";", topics));
-        KafkaSyncDatabaseAction action =
-                syncDatabaseActionBuilder(kafkaConfig)
-                        .withTableConfig(getBasicTableConfig())
-                        .build();
-        runActionWithDefaultEnv(action);
-
-        testSchemaEvolutionImpl(topics, writeOne, fileCount);
+        testSchemaEvolutionMultiTopic(OGG);
     }
 
     @Test
     @Timeout(60)
     public void testSchemaEvolutionOneTopic() throws Exception {
-        final String topic = "schema_evolution";
-        boolean writeOne = true;
-        int fileCount = 2;
-        List<String> topics = Collections.singletonList(topic);
-        topics.forEach(t -> createTestTopic(t, 1, 1));
-
-        // ---------- Write the ogg json into Kafka -------------------
-
-        for (int i = 0; i < fileCount; i++) {
-            try {
-                writeRecordsToKafka(
-                        topics.get(0),
-                        readLines(
-                                "kafka/ogg/database/schemaevolution/topic"
-                                        + i
-                                        + "/ogg-data-1.txt"));
-            } catch (Exception e) {
-                throw new Exception("Failed to write ogg data to Kafka.", e);
-            }
-        }
-
-        Map<String, String> kafkaConfig = getBasicKafkaConfig();
-        kafkaConfig.put(VALUE_FORMAT.key(), "ogg-json");
-        kafkaConfig.put(TOPIC.key(), String.join(";", topics));
-        KafkaSyncDatabaseAction action =
-                syncDatabaseActionBuilder(kafkaConfig)
-                        .withTableConfig(getBasicTableConfig())
-                        .build();
-        runActionWithDefaultEnv(action);
-
-        testSchemaEvolutionImpl(topics, writeOne, fileCount);
-    }
-
-    private void testSchemaEvolutionImpl(List<String> topics, boolean writeOne, int fileCount)
-            throws Exception {
-        waitingTables("T1", "T2");
-
-        FileStoreTable table1 = getFileStoreTable("T1");
-        FileStoreTable table2 = getFileStoreTable("T2");
-
-        RowType rowType1 =
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING().notNull(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING()
-                        },
-                        new String[] {"id", "name", "description", "weight"});
-        List<String> primaryKeys1 = Collections.singletonList("id");
-        List<String> expected =
-                Arrays.asList(
-                        "+I[101, scooter, Small 2-wheel scooter, 3.140000104904175]",
-                        "+I[102, car battery, 12V car battery, 8.100000381469727]");
-        waitForResult(expected, table1, rowType1, primaryKeys1);
-
-        RowType rowType2 =
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING().notNull(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING()
-                        },
-                        new String[] {"id", "name", "description", "weight"});
-        List<String> primaryKeys2 = Collections.singletonList("id");
-        List<String> expected2 =
-                Arrays.asList(
-                        "+I[103, 12-pack drill bits, 12-pack of drill bits with sizes ranging from #40 to #3, 0.800000011920929]",
-                        "+I[104, hammer, 12oz carpenter's hammer, 0.75]");
-        waitForResult(expected2, table2, rowType2, primaryKeys2);
-
-        for (int i = 0; i < fileCount; i++) {
-            try {
-                writeRecordsToKafka(
-                        writeOne ? topics.get(0) : topics.get(i),
-                        readLines(
-                                "kafka/ogg/database/schemaevolution/topic"
-                                        + i
-                                        + "/ogg-data-2.txt"));
-            } catch (Exception e) {
-                throw new Exception("Failed to write ogg data to Kafka.", e);
-            }
-        }
-
-        rowType1 =
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING().notNull(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING()
-                        },
-                        new String[] {"id", "name", "description", "weight", "age"});
-        expected =
-                Arrays.asList(
-                        "+I[101, scooter, Small 2-wheel scooter, 3.140000104904175, NULL]",
-                        "+I[102, car battery, 12V car battery, 8.100000381469727, NULL]",
-                        "+I[103, 12-pack drill bits, 12-pack of drill bits with sizes ranging from #40 to #3, 0.800000011920929, 19]",
-                        "+I[104, hammer, 12oz carpenter's hammer, 0.75, 25]");
-        waitForResult(expected, table1, rowType1, primaryKeys1);
-
-        rowType2 =
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING().notNull(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING()
-                        },
-                        new String[] {"id", "name", "description", "weight", "address"});
-        expected =
-                Arrays.asList(
-                        "+I[103, 12-pack drill bits, 12-pack of drill bits with sizes ranging from #40 to #3, 0.800000011920929, Beijing]",
-                        "+I[104, hammer, 12oz carpenter's hammer, 0.75, Shanghai]");
-        waitForResult(expected, table2, rowType2, primaryKeys2);
+        testSchemaEvolutionOneTopic(OGG);
     }
 
     @Test
     public void testTopicIsEmpty() {
-        Map<String, String> kafkaConfig = getBasicKafkaConfig();
-        kafkaConfig.put(VALUE_FORMAT.key(), "ogg-json");
-
-        KafkaSyncDatabaseAction action = syncDatabaseActionBuilder(kafkaConfig).build();
-
-        assertThatThrownBy(action::run)
-                .satisfies(
-                        AssertionUtils.anyCauseMatches(
-                                IllegalArgumentException.class,
-                                "kafka_conf must and can only set one of the following options: topic,topic-pattern."));
+        testTopicIsEmpty(OGG);
     }
 
     @Test
     @Timeout(60)
     public void testTableAffixMultiTopic() throws Exception {
-        // create table t1
-        createFileStoreTable(
-                "TEST_PREFIX_T1_TEST_SUFFIX",
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING()
-                        },
-                        new String[] {"id", "name", "description", "weight"}),
-                Collections.emptyList(),
-                Collections.singletonList("id"),
-                Collections.emptyMap());
-
-        final String topic1 = "prefix_suffix_0";
-        final String topic2 = "prefix_suffix_1";
-        boolean writeOne = false;
-        int fileCount = 2;
-        List<String> topics = Arrays.asList(topic1, topic2);
-        topics.forEach(topic -> createTestTopic(topic, 1, 1));
-
-        // ---------- Write the ogg json into Kafka -------------------
-
-        for (int i = 0; i < topics.size(); i++) {
-            try {
-                writeRecordsToKafka(
-                        topics.get(i),
-                        readLines("kafka/ogg/database/prefixsuffix/topic" + i + "/ogg-data-1.txt"));
-            } catch (Exception e) {
-                throw new Exception("Failed to write ogg data to Kafka.", e);
-            }
-        }
-
-        // try synchronization
-        Map<String, String> kafkaConfig = getBasicKafkaConfig();
-        kafkaConfig.put(VALUE_FORMAT.key(), "ogg-json");
-        kafkaConfig.put(TOPIC.key(), String.join(";", topics));
-        KafkaSyncDatabaseAction action =
-                syncDatabaseActionBuilder(kafkaConfig)
-                        .withTablePrefix("TEST_PREFIX_")
-                        .withTableSuffix("_TEST_SUFFIX")
-                        .withTableConfig(getBasicTableConfig())
-                        // test including check with affix
-                        .includingTables(ThreadLocalRandom.current().nextBoolean() ? "T1|T2" : ".*")
-                        .build();
-        runActionWithDefaultEnv(action);
-
-        testTableAffixImpl(topics, writeOne, fileCount);
+        testTableAffixMultiTopic(OGG);
     }
 
     @Test
     @Timeout(60)
     public void testTableAffixOneTopic() throws Exception {
-        // create table t1
-        createFileStoreTable(
-                "TEST_PREFIX_T1_TEST_SUFFIX",
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING()
-                        },
-                        new String[] {"id", "name", "description", "weight"}),
-                Collections.emptyList(),
-                Collections.singletonList("id"),
-                Collections.emptyMap());
-
-        final String topic1 = "prefix_suffix";
-        List<String> topics = Collections.singletonList(topic1);
-        boolean writeOne = true;
-        int fileCount = 2;
-        topics.forEach(topic -> createTestTopic(topic, 1, 1));
-
-        // ---------- Write the ogg json into Kafka -------------------
-
-        for (int i = 0; i < fileCount; i++) {
-            try {
-                writeRecordsToKafka(
-                        topics.get(0),
-                        readLines("kafka/ogg/database/prefixsuffix/topic" + i + "/ogg-data-1.txt"));
-            } catch (Exception e) {
-                throw new Exception("Failed to write ogg data to Kafka.", e);
-            }
-        }
-
-        // try synchronization
-        Map<String, String> kafkaConfig = getBasicKafkaConfig();
-        kafkaConfig.put(VALUE_FORMAT.key(), "ogg-json");
-        kafkaConfig.put(TOPIC.key(), String.join(";", topics));
-        KafkaSyncDatabaseAction action =
-                syncDatabaseActionBuilder(kafkaConfig)
-                        .withTablePrefix("TEST_PREFIX_")
-                        .withTableSuffix("_TEST_SUFFIX")
-                        .withTableConfig(getBasicTableConfig())
-                        // test including check with affix
-                        .includingTables(ThreadLocalRandom.current().nextBoolean() ? "T1|T2" : ".*")
-                        .build();
-        runActionWithDefaultEnv(action);
-
-        testTableAffixImpl(topics, writeOne, fileCount);
-    }
-
-    private void testTableAffixImpl(List<String> topics, boolean writeOne, int fileCount)
-            throws Exception {
-        waitingTables("TEST_PREFIX_T1_TEST_SUFFIX", "TEST_PREFIX_T2_TEST_SUFFIX");
-
-        FileStoreTable table1 = getFileStoreTable("TEST_PREFIX_T1_TEST_SUFFIX");
-        FileStoreTable table2 = getFileStoreTable("TEST_PREFIX_T2_TEST_SUFFIX");
-
-        RowType rowType1 =
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING().notNull(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING()
-                        },
-                        new String[] {"id", "name", "description", "weight"});
-        List<String> primaryKeys1 = Collections.singletonList("id");
-        List<String> expected =
-                Arrays.asList(
-                        "+I[101, scooter, Small 2-wheel scooter, 3.140000104904175]",
-                        "+I[102, car battery, 12V car battery, 8.100000381469727]");
-        waitForResult(expected, table1, rowType1, primaryKeys1);
-
-        RowType rowType2 =
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING().notNull(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING()
-                        },
-                        new String[] {"id", "name", "description", "weight"});
-        List<String> primaryKeys2 = Collections.singletonList("id");
-        expected =
-                Arrays.asList(
-                        "+I[103, 12-pack drill bits, 12-pack of drill bits with sizes ranging from #40 to #3, 0.800000011920929]",
-                        "+I[104, hammer, 12oz carpenter's hammer, 0.75]");
-        waitForResult(expected, table2, rowType2, primaryKeys2);
-
-        for (int i = 0; i < fileCount; i++) {
-            try {
-                writeRecordsToKafka(
-                        writeOne ? topics.get(0) : topics.get(i),
-                        readLines("kafka/ogg/database/prefixsuffix/topic" + i + "/ogg-data-2.txt"));
-            } catch (Exception e) {
-                throw new Exception("Failed to write ogg data to Kafka.", e);
-            }
-        }
-        rowType1 =
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING().notNull(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING()
-                        },
-                        new String[] {"id", "name", "description", "weight", "address"});
-        expected =
-                Arrays.asList(
-                        "+I[101, scooter, Small 2-wheel scooter, 3.140000104904175, Beijing]",
-                        "+I[102, car battery, 12V car battery, 8.100000381469727, Shanghai]");
-        waitForResult(expected, table1, rowType1, primaryKeys1);
-
-        rowType2 =
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING().notNull(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING()
-                        },
-                        new String[] {"id", "name", "description", "weight", "age"});
-        expected =
-                Arrays.asList(
-                        "+I[103, 12-pack drill bits, 12-pack of drill bits with sizes ranging from #40 to #3, 0.800000011920929, 19]",
-                        "+I[104, hammer, 12oz carpenter's hammer, 0.75, 25]");
-        waitForResult(expected, table2, rowType2, primaryKeys2);
+        testTableAffixOneTopic(OGG);
     }
 
     @Test
     @Timeout(60)
     public void testIncludingTables() throws Exception {
-        includingAndExcludingTablesImpl(
-                "FLINK|PAIMON.+",
-                null,
-                Arrays.asList("FLINK", "PAIMON_1", "PAIMON_2"),
-                Collections.singletonList("IGNORE"));
+        testIncludingTables(OGG);
     }
 
     @Test
     @Timeout(60)
     public void testExcludingTables() throws Exception {
-        includingAndExcludingTablesImpl(
-                null,
-                "FLINK|PAIMON.+",
-                Collections.singletonList("IGNORE"),
-                Arrays.asList("FLINK", "PAIMON_1", "PAIMON_2"));
+        testExcludingTables(OGG);
     }
 
     @Test
     @Timeout(60)
     public void testIncludingAndExcludingTables() throws Exception {
-        includingAndExcludingTablesImpl(
-                "FLINK|PAIMON.+",
-                "PAIMON_1",
-                Arrays.asList("FLINK", "PAIMON_2"),
-                Arrays.asList("PAIMON_1", "IGNORE"));
-    }
-
-    private void includingAndExcludingTablesImpl(
-            @Nullable String includingTables,
-            @Nullable String excludingTables,
-            List<String> existedTables,
-            List<String> notExistedTables)
-            throws Exception {
-        final String topic1 = "include_exclude" + UUID.randomUUID();
-        List<String> topics = Collections.singletonList(topic1);
-        topics.forEach(topic -> createTestTopic(topic, 1, 1));
-
-        // ---------- Write the ogg json into Kafka -------------------
-
-        try {
-            writeRecordsToKafka(
-                    topics.get(0), readLines("kafka/ogg/database/include/topic0/ogg-data-1.txt"));
-        } catch (Exception e) {
-            throw new Exception("Failed to write ogg data to Kafka.", e);
-        }
-        // try synchronization
-        Map<String, String> kafkaConfig = getBasicKafkaConfig();
-        kafkaConfig.put(VALUE_FORMAT.key(), "ogg-json");
-        kafkaConfig.put(TOPIC.key(), String.join(";", topics));
-        KafkaSyncDatabaseAction action =
-                syncDatabaseActionBuilder(kafkaConfig)
-                        .includingTables(includingTables)
-                        .excludingTables(excludingTables)
-                        .withTableConfig(getBasicTableConfig())
-                        .build();
-        runActionWithDefaultEnv(action);
-
-        // check paimon tables
-        waitingTables(existedTables);
-        assertTableNotExists(notExistedTables);
+        testIncludingAndExcludingTables(OGG);
     }
 
     @Test
     @Timeout(60)
     public void testCaseInsensitive() throws Exception {
-        final String topic = "case-insensitive";
-        createTestTopic(topic, 1, 1);
-
-        // ---------- Write the ogg json into Kafka -------------------
-
-        writeRecordsToKafka(topic, readLines("kafka/ogg/database/case-insensitive/ogg-data-1.txt"));
-
-        Map<String, String> kafkaConfig = getBasicKafkaConfig();
-        kafkaConfig.put(VALUE_FORMAT.key(), "ogg-json");
-        kafkaConfig.put(TOPIC.key(), topic);
-
-        KafkaSyncDatabaseAction action =
-                syncDatabaseActionBuilder(kafkaConfig)
-                        .withTableConfig(getBasicTableConfig())
-                        .withCatalogConfig(
-                                Collections.singletonMap(
-                                        FileSystemCatalogOptions.CASE_SENSITIVE.key(), "false"))
-                        .build();
-        runActionWithDefaultEnv(action);
-
-        waitingTables("t1");
-        FileStoreTable table = getFileStoreTable("t1");
-        RowType rowType =
-                RowType.of(
-                        new DataType[] {
-                            DataTypes.STRING().notNull(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING(),
-                            DataTypes.STRING()
-                        },
-                        new String[] {"id", "name", "description", "weight"});
-        List<String> primaryKeys1 = Collections.singletonList("id");
-        List<String> expected =
-                Arrays.asList(
-                        "+I[101, scooter, Small 2-wheel scooter, 3.140000104904175]",
-                        "+I[102, car battery, 12V car battery, 8.100000381469727]");
-        waitForResult(expected, table, rowType, primaryKeys1);
+        testCaseInsensitive(OGG);
     }
 }
