+++ paimon-flink/paimon-flink-cdc/src/test/resources/kafka/ogg/database/case-insensitive/ogg-data-1.txt
@@ -17,5 +17,5 @@
  */
 
 
-{"table":"PAIMON_SYNC_DATABASE.T1","pos":"00000000000000000000143","primary_keys":["ID"],"after":{"ID":101,"NAME":"scooter","description":"Small 2-wheel scooter","weight":3.140000104904175},"op_type":"I", "current_ts":"2020-05-13T13:39:35.766000", "op_ts":"2020-05-13 15:40:06.000000"}
-{"table":"PAIMON_SYNC_DATABASE.T1","pos":"00000000000000000000144","primary_keys":["id"],"after":{"ID":102,"name":"car battery","description":"12V car battery","WEIGHT":8.100000381469727},"op_type":"I","op_ts":"2020-05-13 15:40:07.000000"}
+{"table":"PAIMON_SYNC_DATABASE.T1","pos":"00000000000000000000143","primary_keys":["ID"],"after":{"ID":101,"NAME":"scooter","description":"Small 2-wheel scooter","weight":3.14},"op_type":"I", "current_ts":"2020-05-13T13:39:35.766000", "op_ts":"2020-05-13 15:40:06.000000"}
+{"table":"PAIMON_SYNC_DATABASE.T1","pos":"00000000000000000000144","primary_keys":["id"],"after":{"ID":102,"name":"car battery","description":"12V car battery","WEIGHT":8.1},"op_type":"I","op_ts":"2020-05-13 15:40:07.000000"}
