+++ paimon-flink/paimon-flink-cdc/src/test/resources/kafka/ogg/database/schemaevolution/topic1/ogg-data-2.txt
@@ -16,5 +16,5 @@
  * limitations under the License.
  */
 
-{"table":"PAIMON_SYNC_DATABASE.T2","pos":"00000000000000000000145","primary_keys":["id"],"after":{"id":103,"name":"12-pack drill bits","description":"12-pack of drill bits with sizes ranging from #40 to #3","weight":0.800000011920929,"address":"Beijing"},"op_type":"I","op_ts":"2020-05-13 15:40:07.000000"}
-{"table":"PAIMON_SYNC_DATABASE.T2","pos":"00000000000000000000146","primary_keys":["id"],"after":{"id":104,"name":"hammer","description":"12oz carpenter's hammer","weight":0.75,"address":"Shanghai"},"op_type":"I","op_ts":"2020-05-13 15:40:07.000000"}
+{"table":"PAIMON_SYNC_DATABASE.t2","pos":"00000000000000000000145","primary_keys":["id"],"after":{"id":103,"name":"12-pack drill bits","description":"12-pack of drill bits with sizes ranging from #40 to #3","weight":0.8,"address":"Beijing"},"op_type":"I","op_ts":"2020-05-13 15:40:07.000000"}
+{"table":"PAIMON_SYNC_DATABASE.t2","pos":"00000000000000000000146","primary_keys":["id"],"after":{"id":104,"name":"hammer","description":"12oz carpenter's hammer","weight":0.75,"address":"Shanghai"},"op_type":"I","op_ts":"2020-05-13 15:40:07.000000"}
