+++ paimon-flink/paimon-flink-cdc/src/test/resources/kafka/ogg/table/event/event-delete.txt
@@ -16,4 +16,4 @@
  * limitations under the License.
  */
 
-{"table":"OGG.EVENT","pos":"00000000000000000000158","primary_keys":["id"],"before":{"id":103,"name":"scooter","description":"Big 2-wheel scooter ","weight":8.170000076293945},"after":null,"op_type":"D","op_ts":"2020-05-13 17:32:24.455000"}
+{"table":"OGG.EVENT","pos":"00000000000000000000158","primary_keys":["id"],"before":{"id":103,"name":"scooter","description":"Big 2-wheel scooter ","weight":8.1},"after":null,"op_type":"D","op_ts":"2020-05-13 17:32:24.455000"}
