+++ paimon-flink/paimon-flink-cdc/src/test/resources/kafka/ogg/table/event/event-insert.txt
@@ -16,8 +16,8 @@
  * limitations under the License.
  */
 
-{"table":"OGG.EVENT","pos":"00000000000000000000143","primary_keys":["id"],"after":{"id":101,"name":"scooter","description":"Small 2-wheel scooter","weight":3.140000104904175},"op_type":"I", "current_ts":"2020-05-13T13:39:35.766000", "op_ts":"2020-05-13 15:40:06.000000"}
-{"table":"OGG.EVENT","pos":"00000000000000000000144","primary_keys":["id"],"after":{"id":102,"name":"car battery","description":"12V car battery","weight":8.100000381469727},"op_type":"I","op_ts":"2020-05-13 15:40:07.000000"}
-{"table":"OGG.EVENT","pos":"00000000000000000000155","primary_keys":["id"],"after":{"id":103,"name":"scooter","description":"Big 2-wheel scooter ","weight":5.179999828338623},"op_type":"I","op_ts":"2020-05-13 17:30:43.428000"}
+{"table":"OGG.EVENT","pos":"00000000000000000000143","primary_keys":["id"],"after":{"id":101,"name":"scooter","description":"Small 2-wheel scooter","weight":3.14},"op_type":"I", "current_ts":"2020-05-13T13:39:35.766000", "op_ts":"2020-05-13 15:40:06.000000"}
+{"table":"OGG.EVENT","pos":"00000000000000000000144","primary_keys":["id"],"after":{"id":102,"name":"car battery","description":"12V car battery","weight":8.1},"op_type":"I","op_ts":"2020-05-13 15:40:07.000000"}
+{"table":"OGG.EVENT","pos":"00000000000000000000155","primary_keys":["id"],"after":{"id":103,"name":"scooter","description":"Big 2-wheel scooter ","weight":5.1},"op_type":"I","op_ts":"2020-05-13 17:30:43.428000"}
